refactor: extract studio MCP tools into modular files
Extract 13 existing MCP tools from monolithic studio-mcp.service.ts
(727 lines) into separate files under tools/ directory:

- tools/types.ts: shared types, permission checker, result helpers
- tools/workflow.tools.ts: 7 workflow tools + task monitor
- tools/component.tools.ts: 2 component tools
- tools/run.tools.ts: 4 run tools

Service is now a thin orchestrator (~55 lines) that creates the
MCP server and delegates to tool registration functions.
